backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
Pillow==11.2.1
waitress==2.1.2
pyinstaller==6.5.0
gunicorn==21.2.0
